Don't treat "." as valid float (#50)
When encountering a decimal character and then immediately an EOF, we
were missing a check that we actually parsed any other digits and just
assumed parsing succeeded; this led to parsing "." as `0.0`. This PR
adds a check that if we saw decimal char and then EOF, if the decimal
char was the only byte consumed, then the value is invalid.