Fix issue parsing AM/PM for dates
In Parsers.jl, we define our own core date parsing routines that mirror
the stdlib's, except ours work on `AbstractVector{UInt8}`. In this case,
the stdlib method for `tryparsenext` on a `p` DatePart had an assertion
that what they were iterating was a `Char`, when in Parsers.jl it's a
`UInt8`. We fix this by overloading the stdlib definition and defining
it for `AbstractVector{UInt8}`.