Ensure invalid dates follow Parsers error flow
This came up in a CSV.jl issue:
https://github.com/JuliaData/CSV.jl/issues/86#issuecomment-624022339,
where an error was being thrown while parsing due to an invalid date.
This breaks the error flow of CSV & Parsers which have their own
error handling stack and signal and handle invalid values without
throwing errors. In the automatic inference case of CSV.jl (where it
decides the types of columns itself while parsing), this should simply
promote the column to a String type. In Parsers with `Parsers.parse`, trying to parse an
invalid date should follow the same pattern as other types and throw a
`Parsers.Error` with the appropriate error codes set on the return code
instead of allowing the Dates module to throw its own error.