Don't parse more digits that we know we'll need
Originally posted as a security issue here:
https://github.com/quinnj/JSON3.jl/issues/57. The problem was that no
matter how many digits were in the input string, we tried to consume
them all, even though there are limits to the number of digits that can
realistically be useful for a given float precision. The change here
checks for those limits and if we go beyond the allowed # of digits,
we'll mark the input as invalid and bail; thus avoiding the "hang"
encountered in the original issue.