Fix successful parsing of single decimal point as valid float
Fixes #53. The issue here is for a single decimal point that was
followed by a delimiter, we weren't hitting the normal code paths like
EOF that marked this invalid. We use the `ndigits` variable that tracks
how many digits we've parsed to check to ensure we've parsed _some_
digits before marking INVALID.