Make sure Parsers.parse consume full input for non-IO arguments
There were cases like `"10a"` where `Parsers.parse` would consume the
`10` and then stop parsing because it encountered `a` and return
successful. We now have a check for non-IO arguments that we consumed
the full input. For IO, we keep the same behavior where digits will be
consumed and we'll stop once we encountered non-digit. Fixes #55.