Fix edge cases of floats on various type promotion boundaries (#63)
Fixes JuliaData/CSV.jl#710.

The issue here is that we weren't accounting for exp adjustments when
considering various boundaries. That is, the negative exp boundary was
325, but because you could have a number like 4.94e-324, the exp
actually gets adjusted to 326 (because of the 2 digits after the decimal
place), and we can still accurately calculate this number, but our
boundary was set at 325. This PR adjusts the boundaries to account for
these adjustments and allows sucessfully parsing some of the smallest
allowed floats.