Fix signed integer overflow case for float parsing
Fixes https://github.com/JuliaData/CSV.jl/issues/769. The issue here is
our `digits` variable for tracking all digits we've parsed in decimal
form is an unsigned integer, but might be converted to a signed integer
in cases where there's no decimal point or a decimal point with no
trailing digits. In those cases, there's an edge case of overflow
because the overflow check was on the _unsigned_ integer instead of the
_signed_ integer, which led to the conversion error. The fix is to
instead change the overflow check to be on the _signed_ integer.