Make a copy of input sentinels if provided (#73)
Fixes https://github.com/JuliaData/CSV.jl/issues/780. The issue here is
Parsers.Options was assuming ownership of and modifying the input
`sentinel` if it was a `Vector{String}`. The fix is to make a copy
instead and then we can modify and use it to store references to other
strings as needed.