Fix undefined variable error when sentinel is provided but we reach EOF (#77)
Reported by Nils Gudat on slack. The problem here is that we defined our
`sentinelpos` variable at the top of the String parsing method, but
didn't defint `sentstart` until later, since by definition that variable
tracks where the sentinel starts, which might be after an opening quote
character. If we happened to start parsing and reach EOF, with a
provided sentinel string, we then checked at the end of the routine if
`sentstart == pos` and got an undefined variable error. The fix is
pretty easy, just define `sentstart` to be 0 at the top of the routine
along with `sentinelpos` to ensure it's always defined.