Fix timetype parsing on IO sources (#79)
Fixes #60. This is an annoying bug because the stdlib Dates parsing
machinery is all String-based, and the Parsers Dates parsing machinery
is all AbstractVector{UInt8}-based, which left IO out to dry. And
unfortunately, too annoying to write at the most performance low-level.
So we take the easier, but at least correct way out. We parse the date
string as a String, then pass the parsed string to the bytes-based
parsing machinery. So def a little slower than if you're parsing
directly from a string or byte vector, but not crazy bad.