Provide getstring utility function for getting parsed string (#80)
Fixed #35. Pretty simple/straightforward. If you want the actual parsed
string, then pass your source + vpos + vlen to `Parsers.getstring` to
get the parsed string out. `vpos` and `vlen` are two of the return
values from `Parsers.xparse` which represent the starting byte position
of the value (`vpos`) and the length in bytes of the parsed string
(`vlen`).