Ensure all BigInts are initialized at __init__-time (#89)
Attempt to fix #87. My theory for what's going on here is that because
the `BIGINT` array is a global const like `const BIGINT = [BigInt()]`,
the internal pointers of the 1st `BigInt` element are becoming stale
once it goes through PackageCompiler. The proposed change here instead
sets the `BIGINT` global const as an _empty_ array, and then intializes
_all_ elements per thread during module `__init__` function.