Ensure high-precision floats don't default to Float64 for scaling (#91)
The issue here is that in the `_scale` method where `v::UInt128`, `v`
might be so large as to lose precision when doing the `x = v /
exp10(-exp)` operation since `exp10(-exp)` will be a `Float64` and hence
our `v` will be promoted to `Float64` before dividing. That's fine if `v
< maxsig(T)`, like in the original `scale` method, but otherwise, we
need to widen `v` and do the division via `BigInt`.