Allow setting of parsed value if typeparser succeeded
As pointed out by @nickrobinson251, if calling `typeparser` succeeds
(i.e. `(code & OK) == OK`), then we might as well set the correct value
in the parsing `Result` object. Up till now, if there was any `INVALID`
code invovled, `res.val` was undefined. With the proposed change in this
PR, we introduce a `Parsers.valueok(code)` function that can be checked,
and if true, then the user can know that a valid value was parsed and
can be accessed via `res.val`. Closes #93.