Fix case where we have a dangling non-digit after integer in float parse (#103)
Fixes issue reported on slack. The problem is that we have a code path
in `parsedigits` that exits early if we hit EOF and just converts the
integer string to the desired float type. But if there's a dangling
non-digit, like "123A", then we parse the "123" but then there may be
scientific notation or a decimal point, so we pass the parsing off to
the `parsefrac` function which handles both those cases. `parsefrac`
however, doesn't have the convert-integer-to-float return option, so we
pass the parsed digits and, in this case non-existent, `exp` number to
`scale` to "scale" the digits appropriately for the float case. `scale`
can normally handle an `exp` of `0`, _except_ in the case where the
`digits` are bigger than the `maxsig` allowed for `Float64`, in which
case it's converted to a `UInt128` and the `BigInt`/`BigFloat` code
takes care of the scaling. That's where the originally reported error
was happening because those "big" code paths weren't expecting an `exp`
of 0.

The solution proposed here is to avoid getting into the `scale` business
at all, since it's 1) more efficient and 2) we're basically still in the
state at the end of `parsedigits` where we've basically just parsed an
integer and can just convert it directly to the appropriate float type.