Fix corner cases with stripewhitescpace and quoted/delimited strings (#107)
Fixes #106. There were a couple of corner cases missed in the initial
stripwhitespace support for strings when quotes are involved. In
particular, if a string is quoted, then we want to remember the last
non-whitespace position _inside_ the quotes and _not_ update the
position outside the quotes. There were also a few invalid cases we
weren't marking correctly if non-whitespace-non-delimiter characters
were encountered in between close quote character and delimiter.