Fix corner case of very large integer parsing that promotes to BigInt (#118)
Fixes https://github.com/JuliaData/CSV.jl/issues/1007. The issue here is there's
actually a git sha "19129370688824811353f0bcee35b917" where the multithreaded
CSV.File case was trying to detect if this value was a float or not. But attempting
to call `Parsers.xparse` threw an error, which shouldn't happen from the Parsers.xparse
code path (it should just return an error code in the parsing result). The problem
was it parsed the very large integer 19129370688824811353, then `f` which it treated
as the exponenent character, then `0` as the exponent. When it went to do the scaling
code, it promoted the integer to `BigInt`, but the `_scale` BigInt path didn't
handle `0` exponent form. This is easily handled before we promote to `BigInt`
since we're really just looking to convert the large integer to the desired type.