Treat empty quoted field as not empty sentinel for Strings
Implements https://github.com/JuliaData/Parsers.jl/issues/138.

The biggest question I have is whether we should view this as a bugfix
or a breaking change in behavior. I think it's arguable either way.

As a note, this doesn't break anything in current CSV.jl code,
which is maybe a signal that this change in behavior isn't super
common or makes that much of a difference?

cc: @nickrobinson251, @Drvi