Ensure Char doesn't match provided delim before marking OK
In the switch to making Char non-greedy in the 2.5 release, I
failed to account for the fact that the parsed Char may in fact
be the delimiter provided, in which case we want to match the delim
instead of treating it as a successfully parsed Char (Char is greedy
in this way like Strings).

Luckily we can still do this pretty efficiently by checking our parsed
Char against the delimiter, if there is one.