Fix a few corner cases of new Number parsing (#161)
If we encounter a decimal point, ensure we treat the `Number` as float
instead of integer. For the case where we were _almost_ overflowing,
we eagerly widen, but if we don't actually overflow, we want to "unwiden"
back to the next lowest integer type. Also fix a few String parsing cases.

At some point, we'll make parsing from a string more first-class, but
for now, it works when calling the Parsers.parsenumber directly.